pyproject.toml
[project]
name = "pipeline-api-agno"
version = "0.1.0"
description = ""
authors = [
    {name = "sch-paulo",email = "pschmitzjr@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13,<4.0"
dependencies = [
    "requests (>=2.32.3,<3.0.0)",
    "tinydb (>=4.8.2,<5.0.0)",
    "sqlalchemy (>=2.0.40,<3.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "streamlit (>=1.45.1,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
